Extend FlagMega model coverage, packing propagation and fusion infrastructure
Add Qwen3.5 MoE/GDN primitives and tutorial reproduction; optimize distribution, readonly materialization and bufferization; improve readable IR dumps; add partial RoPE and Pattern-based QKV/distributed fusion rules with transactional region rewrites.

Keep the TLE-guarded scalar reduction encoding fix and its regressions. Validation: 1184 focused CPU tests and 6 native GPU regressions pass; current Qwen3-1.7B 28-layer GPU and serving performance/greedy-token checks pass. Existing function-boundary topology and full-model call-frame-equality test failures remain documented locally and are not claimed fixed.

+++ python/test/flagmega/ir/ops/nn/test_rope.py
@@ -32,7 +32,7 @@ def test_rope_broadcasts_position_pair_and_rotates_half_dimensions():
         module, {"value": value, "cos": cos, "sin": sin})[0]
 
     rotated = torch.cat((-value[..., 2:], value[..., :2]), dim=-1)
-    expected = value * cos.to(torch.bfloat16) + rotated * sin.to(torch.bfloat16)
+    expected = (value.float() * cos.float() + rotated.float() * sin.float()).to(value.dtype)
     torch.testing.assert_close(result, expected)
     assert result.dtype == torch.bfloat16
     assert pm.try_match_root(
